+++ models/Document.php
@@ -468,16 +468,18 @@ class Document extends Element\AbstractElement
             $this->setIndex($this->getDao()->getNextIndex());
         }
 
-        // save properties
-        $this->getProperties();
-        $this->getDao()->deleteAllProperties();
-        foreach ($this->getProperties() as $property) {
-            if (!$property->getInherited()) {
-                $property->setDao(null);
-                $property->setCid($this->getId());
-                $property->setCtype('document');
-                $property->setCpath($this->getRealFullPath());
-                $property->save();
+        if ($this->isFieldDirty('properties')) {
+            // save properties
+            $properties = $this->getProperties();
+            $this->getDao()->deleteAllProperties();
+            foreach ($properties as $property) {
+                if (!$property->getInherited()) {
+                    $property->setDao(null);
+                    $property->setCid($this->getId());
+                    $property->setCtype('document');
+                    $property->setCpath($this->getRealFullPath());
+                    $property->save();
+                }
             }
         }
 
@@ -488,7 +490,7 @@ class Document extends Element\AbstractElement
 
         foreach ($this->resolveDependencies() as $requirement) {
             if ($requirement['id'] == $this->getId() && $requirement['type'] == 'document') {
-                // dont't add a reference to yourself
+                // don't add a reference to yourself
                 continue;
             } else {
                 $d->addRequirement((int)$requirement['id'], $requirement['type']);
